added: LAB4

+++ LAB4/baitap2/premium.cpp
@@ -1,7 +1,9 @@
 #include "premium.h"
 
-Premium::Premium(int dem, int pv): Phong(dem, "Premium"), PhiPhucVu(pv){}
+// Phuong thuc thiet lap co tham so cua lop Premium
+Premium::Premium(int dem, long long pv): Phong(dem, "Premium"), PhiPhucVu(pv){}
 
+// Phuong thuc tinh doanh thu cua lop Premium
 long long Premium::DoanhThu(){
   return SoDem*500000 + PhiPhucVu;
 }
